restructure components a little
